Issue 13 - Add raw-ish hex float mode
On the 'a' format from C99:

"""
The double argument is rounded and converted to hexadecimal notation in
the style [-]0xh.hhhp[±]d...  If the precision is missing, it is taken
as enough to represent the floating-point number exactly, and no
rounding occurs.
"""

Tested math.sh on Linux (Fedora 22) and FreeBSD CURRENT-ish.