initial support for unpack with issue #35
add a new test for the new struct/assoc format that works on linux, it
looks a *lot* cleaner than before, but the struct command still needs a
lot of work. Note that pack still needs assoc support, but this is easy.

also need a sizeof command so you find the structure size for malloc.